Added: Tilemap, movement animations, dash effects
Finished with the movement code for now I think. Playermove not ready to be deleted yet as it has other code that I still need but otherwise good. Tilemap has also been set up, has made me realise that there are more sprites i will need to make. Also found a bug with the dash ability, not sure if it will still be there in build but essentially if a player presses the dash key repeatedly then the dash happens twice and the player's move speed is reset to the speed it should be when dashing.